refactor: 전역적으로 사용되는 패키지를 common으로 이동
