+++ src/main/java/gift/common/annotation/LoginUser.java
@@ -1,4 +1,4 @@
-package gift.annotation;
+package gift.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
