+++ src/main/java/gift/common/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.common.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ProblemDetail;
