+++ src/main/java/gift/common/interceptor/LoggingInterceptor.java
@@ -1,4 +1,4 @@
-package gift.interceptor;
+package gift.common.interceptor;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
