+++ src/main/java/gift/common/validation/annotation/RestrictedKeyword.java
@@ -1,6 +1,6 @@
-package gift.validation.annotation;
+package gift.common.validation.annotation;
 
-import gift.validation.validator.RestrictedKeywordValidator;
+import gift.common.validation.validator.RestrictedKeywordValidator;
 import jakarta.validation.Constraint;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
