+++ src/main/java/gift/config/WebConfig.java
@@ -1,6 +1,6 @@
 package gift.config;
 
-import gift.resolver.LoginUserArgumentResolver;
+import gift.common.resolver.LoginUserArgumentResolver;
 import java.util.List;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
