+++ src/main/java/gift/product/dto/request/CreateProductRequest.java
@@ -1,6 +1,6 @@
 package gift.product.dto.request;
 
-import gift.validation.annotation.RestrictedKeyword;
+import gift.common.validation.annotation.RestrictedKeyword;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotEmpty;
 import jakarta.validation.constraints.NotNull;
