+++ src/main/java/gift/user/service/UserService.java
@@ -1,7 +1,7 @@
 package gift.user.service;
 
-import gift.exception.CustomException;
-import gift.exception.ErrorCode;
+import gift.common.exception.CustomException;
+import gift.common.exception.ErrorCode;
 import gift.user.dto.UserInfo;
 import gift.user.dto.request.UserLoginRequest;
 import gift.user.dto.request.UserRegisterRequest;
