+++ src/main/java/gift/wish/service/WishService.java
@@ -1,7 +1,7 @@
 package gift.wish.service;
 
-import gift.exception.CustomException;
-import gift.exception.ErrorCode;
+import gift.common.exception.CustomException;
+import gift.common.exception.ErrorCode;
 import gift.product.entity.Product;
 import gift.product.repository.ProductJpaRepository;
 import gift.user.entity.User;
