+++ src/test/java/gift/jpa/repository/ProductJpaRepositoryTest.java
@@ -2,8 +2,8 @@ package gift.jpa.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import gift.exception.CustomException;
-import gift.exception.ErrorCode;
+import gift.common.exception.CustomException;
+import gift.common.exception.ErrorCode;
 import gift.product.category.entity.Category;
 import gift.product.category.repository.CategoryJpaRepository;
 import gift.product.entity.Product;
